feat: add AllExceptionFilter, replacing PrismaFilter
