+++ src/all-exception.filter.spec.ts
@@ -1,0 +1,7 @@
+import { AllExceptionFilter } from './all-exception.filter';
+
+describe('AllExceptionFilter', () => {
+  it('should be defined', () => {
+    expect(new AllExceptionFilter()).toBeDefined();
+  });
+});
