--- src/database/prisma.filter.spec.ts
@@ -1,7 +1,0 @@
-import { PrismaFilter } from './prisma.filter';
-
-describe('PrismaFilter', () => {
-  it('should be defined', () => {
-    expect(new PrismaFilter()).toBeDefined();
-  });
-});
